refactor: move channels_v2/stages package to channels/stages
