+++ apps/channels/channels_v2/api_channel.py
@@ -6,7 +6,9 @@ from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext, MessageProcessingPipeline
-from apps.channels.channels_v2.stages.core import (
+from apps.channels.const import MESSAGE_TYPES
+from apps.channels.sender import ChannelSender
+from apps.channels.stages.core import (
     BotInteractionStage,
     ChatMessageCreationStage,
     ConsentCheckStage,
@@ -19,9 +21,7 @@ from apps.channels.channels_v2.stages.core import (
     SessionActivationStage,
     SessionResolutionStage,
 )
-from apps.channels.channels_v2.stages.terminal import ActivityTrackingStage, PersistenceStage
-from apps.channels.const import MESSAGE_TYPES
-from apps.channels.sender import ChannelSender
+from apps.channels.stages.terminal import ActivityTrackingStage, PersistenceStage
 from apps.chat.exceptions import ChannelException
 from apps.chat.models import Chat
 from apps.experiments.models import Experiment, Participant, SessionStatus
