+++ apps/channels/channels_v2/channel_base.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, ClassVar
 
 from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.pipeline import MessageProcessingContext, MessageProcessingPipeline
-from apps.channels.channels_v2.stages.core import (
+from apps.channels.stages.core import (
     AttachmentHydrationStage,
     BotInteractionStage,
     ChatMessageCreationStage,
@@ -20,7 +20,7 @@ from apps.channels.channels_v2.stages.core import (
     SessionActivationStage,
     SessionResolutionStage,
 )
-from apps.channels.channels_v2.stages.terminal import (
+from apps.channels.stages.terminal import (
     ActivityTrackingStage,
     DeliveryErrorHandler,
     PersistenceStage,
