+++ apps/channels/channels_v2/evaluation_channel.py
@@ -7,16 +7,16 @@ from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.api_channel import NoOpSender
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext, MessageProcessingPipeline
-from apps.channels.channels_v2.stages.core import (
+from apps.channels.const import MESSAGE_TYPES
+from apps.channels.sender import ChannelSender
+from apps.channels.stages.core import (
     ChatMessageCreationStage,
     EvalsBotInteractionStage,
     MessageTypeValidationStage,
     QueryExtractionStage,
     ResponseFormattingStage,
 )
-from apps.channels.channels_v2.stages.terminal import ActivityTrackingStage, PersistenceStage
-from apps.channels.const import MESSAGE_TYPES
-from apps.channels.sender import ChannelSender
+from apps.channels.stages.terminal import ActivityTrackingStage, PersistenceStage
 from apps.chat.exceptions import ChannelException
 from apps.service_providers.tracing import TracingService
