+++ apps/channels/channels_v2/telegram_channel.py
@@ -15,10 +15,10 @@ from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities, PlatformConsentConfig
 from apps.channels.channels_v2.channel_base import ChannelBase
 from apps.channels.channels_v2.pipeline import MessageProcessingContext
-from apps.channels.channels_v2.stages.terminal import DeliveryErrorHandler, MessageDeliveryFailure
 from apps.channels.const import MESSAGE_TYPES
 from apps.channels.datamodels import TelegramMessage
 from apps.channels.sender import ChannelSender
+from apps.channels.stages.terminal import DeliveryErrorHandler, MessageDeliveryFailure
 from apps.experiments.models import ParticipantData
 from apps.service_providers.file_limits import can_send_on_telegram
