+++ apps/channels/tests/channels/stages/test_consent_check.py
@@ -3,9 +3,9 @@ from unittest.mock import MagicMock
 import pytest
 
 from apps.channels.capabilities import PlatformConsentConfig
-from apps.channels.channels_v2.stages.core import ConsentCheckStage
 from apps.channels.exceptions import EarlyAbort
 from apps.channels.models import ChannelPlatform
+from apps.channels.stages.core import ConsentCheckStage
 from apps.channels.tests.channels.conftest import make_capabilities, make_context
 from apps.experiments.models import ParticipantData
 from apps.utils.factories.experiment import ParticipantFactory
