+++ apps/channels/tests/channels/stages/test_session_resolution.py
@@ -3,9 +3,9 @@ from unittest.mock import MagicMock
 import pytest
 
 from apps.channels.capabilities import ChannelCapabilities, PlatformConsentConfig
-from apps.channels.channels_v2.stages.core import ConsentCheckStage, SessionResolutionStage
 from apps.channels.const import MESSAGE_TYPES
 from apps.channels.exceptions import EarlyAbort, EarlyExitResponse
+from apps.channels.stages.core import ConsentCheckStage, SessionResolutionStage
 from apps.channels.tests.channels.conftest import make_context
 from apps.channels.tests.message_examples.base_messages import text_message
 from apps.chat.const import STATUSES_FOR_COMPLETE_CHATS
