+++ apps/channels/tests/test_facebook_integration.py
@@ -56,7 +56,7 @@ class TestTwilio:
     @override_settings(WHATSAPP_S3_AUDIO_BUCKET="123")
     @patch("apps.channels.tasks.validate_twillio_request", Mock())
     @patch("apps.service_providers.speech_service.SpeechService.synthesize_voice")
-    @patch("apps.channels.channels_v2.stages.core.QueryExtractionStage._transcribe_voice")
+    @patch("apps.channels.stages.core.QueryExtractionStage._transcribe_voice")
     @patch("apps.service_providers.messaging_service.TwilioService.send_voice_message")
     @patch("apps.service_providers.messaging_service.TwilioService.send_text_message")
     @patch("apps.chat.bots.PipelineBot.process_input")
